genai_rca_assistant/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pyodbc==5.0.1

# Azure integrations
azure-storage-blob==12.19.0

# AI/ML
google-generativeai==0.3.1

# HTTP requests
requests==2.31.0

# WebSocket support
websockets==12.0

# Authentication & Security
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
bcrypt==4.1.1

# Email validation
email-validator==2.1.0

# CORS support (included in FastAPI middleware)